+++ Client/external/GadgetToJScript/.gitignore
