+++ Client/external/GadgetToJScript/build.bat
@@ -1,5 +1,5 @@
 @echo off
-GadgetToJScript.exe -c test.cs -w hta -o test
-GadgetToJScript.exe -c test.cs -w js -o test
-GadgetToJScript.exe -c test.cs -w vba -o test
-GadgetToJScript.exe -c test.cs -w vbs -o test
+GadgetToJScript.exe -c GadgetToJScript.cs -w hta -o test
+GadgetToJScript.exe -c GadgetToJScript.cs -w js -o test
+GadgetToJScript.exe -c GadgetToJScript.cs -w vba -o test
+GadgetToJScript.exe -c GadgetToJScript.cs -w vbs -o test
