+++ Client/gtojs.bat
@@ -2,7 +2,7 @@
 cd external\GadgetToJScript
 call build.bat
 cd ..\..
-cp external\GadgetToJScript\test.hta payload\GadgetToJScript\test.hta
-cp external\GadgetToJScript\test.js payload\GadgetToJScript\test.js
-cp external\GadgetToJScript\test.vba payload\GadgetToJScript\test.vba
-cp external\GadgetToJScript\test.vbs payload\GadgetToJScript\test.vbs
+copy external\GadgetToJScript\test.hta payload\GadgetToJScript\test.hta
+copy external\GadgetToJScript\test.js payload\GadgetToJScript\test.js
+copy external\GadgetToJScript\test.vba payload\GadgetToJScript\test.vba
+copy external\GadgetToJScript\test.vbs payload\GadgetToJScript\test.vbs
