+++ Client/tools/Invoke-Base64.ps1
@@ -1,4 +1,4 @@
-$filename = "tools\csexec.exe"
+$filename = "payload\myclient.exe"
 $old_bytes = [IO.File]::ReadAllBytes($filename)
 [system.Convert]::toBase64string($old_bytes) | Clip
 [system.Convert]::toBase64string($old_bytes)
